+++ src/endian/stream_reader.hpp
@@ -16,7 +16,7 @@ namespace endian
 /// The stream_reader provides a stream-like interface for reading from a
 /// fixed-size buffer. All complexity regarding endianness is encapsulated.
 template <typename EndianType>
-class stream_reader : public detail::stream<const uint8_t*>
+class stream_reader : public detail::stream<detail::const_stream>
 {
 public:
     /// Creates an endian stream on top of a pre-allocated buffer of the